Improved login functionality Login method moved to actions Added signout method

+++ src/store/actions/stickers.js
@@ -1,14 +1,10 @@
 import { ADD_STICKER, EDIT_STICKER, REMOVE_STICKER } from '../action-types';
-import db from '../../services/db/firebase';
+import db from '../../services/db';
 import uuid from 'uuid/v1';
 
-function setSticker(data) {
-  return { type: EDIT_STICKER, payload: data };
-}
-
 export const editSticker = async (dispatch, uid, id, key, value) => {
-  await db.editSticker(uid, { id, [key]: value });
-  dispatch(setSticker({ id, key, value }));
+  if (uid) await db.editSticker(uid, { id, [key]: value });
+  dispatch({ type: EDIT_STICKER, payload: { id, key, value } });
 };
 
 export const addSticker = async (dispatch, uid) => {
@@ -19,11 +15,11 @@ export const addSticker = async (dispatch, uid) => {
     color: '#fff',
     isPinned: false
   };
-  await db.addSticker(uid, newSticker);
+  if (uid) await db.addSticker(uid, newSticker);
   dispatch({ type: ADD_STICKER, payload: newSticker });
 };
 
 export const removeSticker = async (dispatch, uid, id) => {
-  await db.deleteSticker(uid, id);
+  if (uid) await db.deleteSticker(uid, id);
   dispatch({ type: REMOVE_STICKER, payload: id });
 };
